+++ tests/test_libudev.py
@@ -20,109 +20,13 @@ from __future__ import (print_function, division, unicode_literals,
                         absolute_import)
 
 import re
-from subprocess import check_call
-from tempfile import NamedTemporaryFile
-from xml.etree import ElementTree as etree
+import ctypes
 
 import pytest
 
 from pyudev import _libudev as binding
 
 libudev = binding.libudev
-
-
-# new functions added in specific libudev versions
-UDEV_ADDITIONS = {
-    152: ['udev_new_from_environment'],
-    154: ['udev_device_get_tags_list_entry',
-          'udev_enumerate_add_match_tag',
-          'udev_monitor_filter_add_match_tag'],
-    165: ['udev_device_get_is_initialized',
-          'udev_device_get_usec_since_initialized',
-          'udev_enumerate_add_match_is_initialized'],
-    167: ['udev_get_run_path',
-          'udev_device_get_sysattr_list_entry'],
-    172: ['udev_device_has_tag', 'udev_enumerate_add_match_parent'],
-    }
-
-
-LIBUDEV_H = '/usr/include/libudev.h'
-
-
-def get_libudev_functions():
-    with NamedTemporaryFile() as stream:
-        check_call(['gccxml', LIBUDEV_H, '-fxml={0}'.format(stream.name)])
-        tree = etree.parse(stream.name)
-    names = [f.get('name') for f in tree.findall('.//Function')]
-    return [n for n in names if n.startswith('udev_')]
-
-
-def pytest_generate_tests(metafunc):
-    if 'funcname' in metafunc.funcargnames:
-        for namespace, members in binding.SIGNATURES.items():
-            for funcname in members:
-                full_name = '{0}_{1}'.format(namespace, funcname)
-                metafunc.addcall(param=(namespace, funcname), id=full_name,
-                                 funcargs=dict(funcname=full_name))
-
-
-def pytest_funcarg__signature(request):
-    namespace, name = request.param
-    return binding.SIGNATURES[namespace][name]
-
-
-def pytest_funcarg__argtypes(request):
-    argtypes, _ = request.getfuncargvalue('signature')
-    return argtypes
-
-
-def pytest_funcarg__restype(request):
-    _, restype = request.getfuncargvalue('signature')
-    return restype
-
-
-def pytest_funcarg__errcheck(request):
-    funcname = request.getfuncargvalue('funcname')
-    return binding.ERROR_CHECKERS.get(funcname)
-
-
-@pytest.mark.udev_version('>= {0}'.format(max(UDEV_ADDITIONS)))
-def test_presence(funcname):
-    assert hasattr(libudev, funcname)
-
-
-def test_signatures(funcname, restype, argtypes, errcheck):
-    func = getattr(libudev, funcname, None)
-    if not func:
-        pytest.skip('{0} not available'.format(funcname))
-    assert func.restype == restype
-    if not argtypes:
-        assert not func.argtypes
-    else:
-        assert func.argtypes == argtypes
-    assert func.errcheck == errcheck
-
-
-UDEV_ADDITIONS = {
-    152: ['udev_new_from_environment'],
-    154: ['udev_device_get_tags_list_entry',
-          'udev_enumerate_add_match_tag',
-          'udev_monitor_filter_add_match_tag'],
-    165: ['udev_device_get_is_initialized',
-          'udev_device_get_usec_since_initialized',
-          'udev_enumerate_add_match_is_initialized'],
-    167: ['udev_get_run_path',
-          'udev_device_get_sysattr_list_entry'],
-    172: ['udev_device_has_tag', 'udev_enumerate_add_match_parent'],
-    }
-
-
-@pytest.mark.parametrize(('version', 'missing_functions'),
-                         UDEV_ADDITIONS.items())
-def test_missing_functions(version, missing_functions):
-    if pytest.config.udev_version >= version:
-        pytest.skip('udev too new')
-    assert not any(hasattr(libudev, fn) for fn in missing_functions)
 
 
 WRAPPER_BLACKLIST_PATTERNS = [
@@ -142,25 +46,72 @@ WRAPPER_BLACKLIST_PATTERNS = [
 ]
 
 
-def _is_blacklisted(function_name):
+def _is_blacklisted(function):
     for pattern in WRAPPER_BLACKLIST_PATTERNS:
         if pytest.is_unicode_string(pattern):
-            if function_name == pattern:
+            if function.name == pattern:
                 return True
         else:
-            if pattern.match(function_name):
+            if pattern.match(function.name):
                 return True
     else:
         return False
 
 
-@pytest.mark.coverage
-@pytest.mark.parametrize('function_name', get_libudev_functions())
-def test_is_wrapped(function_name):
-    wrapped_functions = set('{0}_{1}'.format(ns, member)
-                            for ns, members in binding.SIGNATURES.items()
-                            for member in members)
-    if _is_blacklisted(function_name):
-        assert function_name not in wrapped_functions
+def pytest_funcarg__libudev_function(request):
+    """
+    Override ``libudev_function`` to skip tests for blacklisted functions.
+    """
+    function = request.getfuncargvalue('libudev_function')
+    if _is_blacklisted(function):
+        pytest.skip('{0} is not wrapped'.format(function.name))
+    return function
+
+
+FUNDAMENTAL_TYPES = {
+    'int': ctypes.c_int,
+    'char': ctypes.c_char,
+    'long long unsigned int': ctypes.c_ulonglong,
+    'long unsigned int': ctypes.c_ulong,
+    'void': None,
+}
+
+
+def _pointer_to_ctypes(pointer):
+    underlying_type = _to_ctypes(pointer.type)
+    if underlying_type is ctypes.c_char:
+        return ctypes.c_char_p
     else:
-        assert function_name in wrapped_functions
+        return ctypes.POINTER(underlying_type)
+
+
+TYPE_CONVERTER = {
+    'FundamentalType': lambda t: FUNDAMENTAL_TYPES[t.name],
+    'Struct': lambda s: getattr(binding, s.name),
+    'PointerType': _pointer_to_ctypes,
+    # const qualifiers are ignored in ctypes
+    'CvQualifiedType': lambda t: _to_ctypes(t.type),
+    # propagate type defs
+    'Typedef': lambda t: _to_ctypes(t.type),
+}
+
+
+def _to_ctypes(libudev_type):
+    return TYPE_CONVERTER[libudev_type.__class__.__name__](libudev_type)
+
+
+def test_wrapper_signature(libudev_function):
+    wrapped_function = getattr(libudev, libudev_function.name)
+    argtypes = [_to_ctypes(a) for a in libudev_function.arguments]
+    assert wrapped_function.argtypes == argtypes
+    restype = _to_ctypes(libudev_function.return_type)
+    assert wrapped_function.restype == restype
+
+
+def test_error_checker(libudev_function):
+    func_name = libudev_function.name
+    if func_name in binding.ERROR_CHECKERS:
+        func = getattr(libudev, func_name)
+        assert func.errcheck == binding.ERROR_CHECKERS[func_name]
+    else:
+        pytest.skip('{0} has no error checker'.format(func_name))